+++ modules/runner-binaries-syncer/lambdas/runner-binaries-syncer/src/lambda.ts
@@ -1,5 +1,6 @@
 import { handle } from './syncer/handler';
 
+// eslint-disable-next-line
 module.exports.handler = async (event: any, context: any, callback: any): Promise<any> => {
   await handle();
   return callback();