+++ modules/runner-binaries-syncer/runner-binaries-syncer.tf
@@ -14,13 +14,20 @@ resource "aws_lambda_function" "syncer" {
 
   environment {
     variables = {
-      S3_BUCKET_NAME             = aws_s3_bucket.action_dist.id
-      S3_OBJECT_KEY              = local.action_runner_distribution_object_key
-      GITHUB_RUNNER_ARCHITECTURE = var.runner_architecture
+      S3_BUCKET_NAME                          = aws_s3_bucket.action_dist.id
+      S3_OBJECT_KEY                           = local.action_runner_distribution_object_key
+      GITHUB_RUNNER_ARCHITECTURE              = var.runner_architecture
+      GITHUB_RUNNER_ALLOW_PRERELEASE_BINARIES = var.runner_allow_prerelease_binaries
     }
   }
 
   tags = var.tags
+}
+
+resource "aws_cloudwatch_log_group" "syncer" {
+  name              = "/aws/lambda/${aws_lambda_function.syncer.function_name}"
+  retention_in_days = var.logging_retention_in_days
+  tags              = var.tags
 }
 
 resource "aws_iam_role" "syncer_lambda" {
@@ -47,7 +54,9 @@ resource "aws_iam_role_policy" "lambda_logging" {
   name = "${var.environment}-lambda-logging-policy-syncer"
   role = aws_iam_role.syncer_lambda.id
 
-  policy = templatefile("${path.module}/policies/lambda-cloudwatch.json", {})
+  policy = templatefile("${path.module}/policies/lambda-cloudwatch.json", {
+    log_group_arn = aws_cloudwatch_log_group.syncer.arn
+  })
 }
 
 resource "aws_iam_role_policy" "syncer" {